Revert "add citecusco_3"
This reverts commit d5ce16a16f47788b71e77f42624b047aa5117fae.

+++ etl/shared/cite.py
@@ -53,7 +53,7 @@ class TransformStep(PipelineStep):
         #cite_map = {k:v for (k,v) in zip(cite_list, list(range(1, len(cite_list) +1)))}
 
         #df['cite_id'] = df['cite'].map(cite_map)
-        df['cite_id'] = df['CITEID'].astype(int)
+        df['cite_id'] = df['CITEID']
  
         df['cite_slug'] = df['cite'].str.lower()
         df['cite_slug'] = df['cite_slug'].apply(unidecode)
